+++ core/src/builtin_widgets/class.rs
@@ -118,16 +118,50 @@ macro_rules! named_styles_impl {
   };
 }
 
-/// This macro is used to define a class implementation by combining multiple
-/// other class implementations.
+/// Combines multiple class implementations into a single implementation.
+/// This macro takes a list of class implementations and returns a closure
+/// that applies each implementation sequentially to a `Widget`.
 #[macro_export]
-macro_rules! multi_class {
+macro_rules! class_multi_impl {
   ($($class: expr),*) => {
     move |mut w: Widget| {
       $(w = $class(w);)*
       w
     }
   };
+}
+
+/// Creates a vector of class names from a list of class expressions. The
+/// resulting array can be used to apply multiple classes to a widget. The last
+/// class in the list will be applied first, as it is closest to the widget.
+///
+/// # Example
+/// ```
+/// use ribir::prelude::*;
+///
+/// class_names!(RED_BORDER, BLUE_BACKGROUND);
+///
+/// let two_classes = class_array!(RED_BORDER, BLUE_BACKGROUND);
+///
+/// let widget = fn_widget! {
+///   @ $two_classes {
+///     @Container {
+///       size: Size::new(100., 100.),
+///     }
+///   }
+/// };
+/// ```
+#[macro_export]
+macro_rules! class_array {
+  ($($class:expr),* $(,)?) => {{
+    [
+      $(
+        $crate::prelude::DeclareInit
+          ::<Option<$crate::prelude::ClassName>>
+          ::declare_from($class)
+      ),*
+    ]
+  }};
 }
 
 /// A empty class implementation that returns the input widget as is.
@@ -317,6 +351,25 @@ impl<'c> ComposeChild<'c> for Class {
       }
     };
     FnWidget::new(f).into_widget()
+  }
+}
+
+impl<'w, const M: usize> ComposeChild<'w> for [DeclareInit<Option<ClassName>>; M] {
+  type Child = Widget<'w>;
+
+  fn compose_child(this: impl StateWriter<Value = Self>, mut widget: Self::Child) -> Widget<'w> {
+    let this = this
+      .try_into_value()
+      .unwrap_or_else(|_| panic!("Class array only supports stateless."));
+    for cls in this.into_iter().rev() {
+      widget = match cls {
+        DeclareInit::Value(class) => Class { class }.with_child(widget).into_widget(),
+        DeclareInit::Pipe(cls) => FatObj::new(widget)
+          .class(DeclareInit::Pipe(cls))
+          .into_widget(),
+      };
+    }
+    widget
   }
 }
 
@@ -541,6 +594,26 @@ mod tests {
     wnd.draw_frame();
     *w_cls.write() = MARGIN;
     wnd.draw_frame();
+  }
+
+  #[test]
+  fn class_array() {
+    reset_test_env!();
+
+    let mut wnd = TestWindow::new(fn_widget! {
+      let margin_and_clamp = class_array![MARGIN, CLAMP_50];
+      @Providers {
+        providers: smallvec![initd_classes().into_provider()],
+        @ $margin_and_clamp {
+          @Container {
+            size: Size::new(100., 100.),
+          }
+        }
+      }
+    });
+
+    wnd.draw_frame();
+    wnd.assert_root_size(Size::new(70., 70.));
   }
 
   #[test]